code_analyser/requirements.txt
docker
boto3
python-dotenv
supabase
